Wait for Javascript-generated elements
The wait tests for the interaction frame were triggered by HTML elements,
but this means the wait only waits till the HTML is loaded.  The subsequent
actions rely on the Javascript to add additional elements.  Since these were
not yet loaded, a fast test would not find the element.  This commit
modifies the tests to use Javascript-generated elements, to ensure they
really are there.